pyproject.toml
[tool.poetry]
name = "wexler_garage_api"
version = "0.2.0"
description = 'Project for automating the voitixler.com course "API and Postman course" on Python'
authors = ["Bohdan Obruch <bodan1992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
faker = "^40.1.2"
pydantic = "^2.10.5"
lorem = "0.1.1"
python-dotenv = "^1.0.1"
curlify = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
allure-pytest = "^2.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

